# lib/api/openapi.yaml
openapi: "3.1.0"
info:
  title: "HAY2010 Stock API"
  version: "1.0.0"
  description: "API REST pour la gestion des partenaires, documents, stock et affaires"
servers:
  - url: "/api/v1"
    description: "Base path for API v1"
security:
  - ApiKeyAuth: []
paths:
...
  /entrepots:
    get:
      operationId: listWarehouses
      summary: "Lister les entrepôts"
      tags:
        - Entrepots
      security:
        - ApiKeyAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: "Recherche dans nom et code"
        - name: principal
          in: query
          schema:
            type: string
            enum: ["true", "false"]
          description: "Filtrer par statut entrepôt principal"
        - name: sort
          in: query
          schema:
            type: string
            enum: [id_entrepot, code_entrepot, nom_entrepot, ville_entrepot, capacite_totale_unites, date_creation, date_modification]
            default: nom_entrepot
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: "Liste paginée des entrepôts"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WarehousePaginatedResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createWarehouse
      summary: "Créer un entrepôt"
      tags:
        - Entrepots
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseCreate"
      responses:
        "201":
          description: "Entrepôt créé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: "Code entrepôt déjà utilisé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /entrepots/{id}:
    get:
      operationId: getWarehouseById
      summary: "Obtenir un entrepôt par ID"
      tags:
        - Entrepots
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Détails de l'entrepôt"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateWarehouse
      summary: "Mettre à jour un entrepôt"
      tags:
        - Entrepots
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseUpdate"
      responses:
        "200":
          description: "Entrepôt mis à jour"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: "Code entrepôt déjà utilisé"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteWarehouse
      summary: "Désactiver un entrepôt (soft delete)"
      tags:
        - Entrepots
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: "Entrepôt désactivé"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: "API key au format hay2010_sk_live_xxxxxxxxxxxxxxxx"
  responses:
    Unauthorized:
      description: "Clé API invalide ou manquante"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: "Requête invalide"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "Ressource non trouvée"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasMore:
          type: boolean
...
    Warehouse:
      type: object
      properties:
        id_entrepot:
          type: integer
        code_entrepot:
          type: string
        nom_entrepot:
          type: string
        adresse_entrepot:
          type: string
          nullable: true
        ville_entrepot:
          type: string
          nullable: true
        code_postal_entrepot:
          type: string
          nullable: true
        capacite_totale_unites:
          type: integer
          nullable: true
        nom_responsable:
          type: string
          nullable: true
        email_responsable:
          type: string
          nullable: true
        telephone_responsable:
          type: string
          nullable: true
        est_actif:
          type: boolean
        est_entrepot_principal:
          type: boolean
        date_creation:
          type: string
          format: date-time
        date_modification:
          type: string
          format: date-time
    WarehouseCreate:
      type: object
      required:
        - code_entrepot
        - nom_entrepot
      properties:
        code_entrepot:
          type: string
          minLength: 1
          maxLength: 50
        nom_entrepot:
          type: string
          minLength: 1
          maxLength: 255
        adresse_entrepot:
          type: string
          nullable: true
        ville_entrepot:
          type: string
          maxLength: 100
          nullable: true
        code_postal_entrepot:
          type: string
          maxLength: 20
          nullable: true
        capacite_totale_unites:
          type: integer
          nullable: true
        nom_responsable:
          type: string
          maxLength: 255
          nullable: true
        email_responsable:
          type: string
          format: email
          maxLength: 255
          nullable: true
        telephone_responsable:
          type: string
          maxLength: 50
          nullable: true
    WarehouseUpdate:
      type: object
      description: "Tous les champs sont optionnels pour la mise à jour partielle"
      properties:
        code_entrepot:
          type: string
          minLength: 1
          maxLength: 50
        nom_entrepot:
          type: string
          minLength: 1
          maxLength: 255
        adresse_entrepot:
          type: string
          nullable: true
        ville_entrepot:
          type: string
          maxLength: 100
          nullable: true
        code_postal_entrepot:
          type: string
          maxLength: 20
          nullable: true
        capacite_totale_unites:
          type: integer
          nullable: true
        nom_responsable:
          type: string
          maxLength: 255
          nullable: true
        email_responsable:
          type: string
          format: email
          maxLength: 255
          nullable: true
        telephone_responsable:
          type: string
          maxLength: 50
          nullable: true
    WarehousePaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Warehouse"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
            properties:
              message:
                type: string
              code:
                type: string
        timestamp:
          type: string
          format: date-time